+++ quantify_scheduler/device_under_test/mock_setup.py
@@ -146,7 +146,7 @@ def set_standard_params_transmon(mock_setup):
     q4.measure.acq_delay(100e-9)
 
 
-def set_up_basic_mock_nv_setup() -> Dict:
+def set_up_mock_basic_nv_setup() -> Dict:
     """Sets up a system containing 1 electronic qubit in an NV center.
 
     After usage, close all instruments.